+++ CHANGELOG.md
@@ -38,6 +38,11 @@
   
 * `integrate/scarches` and `workflows/annotate/scanvi_scarches`: Enable correction for technical variability by multiple continuous and categorical covariates.
 
+
+## BUG FIXES
+
+* `filter/filter_with_counts`: this component would sometimes crash (segfault) when processing malformatted sparse matrices. A proper error message is now provided in this case (PR #1086).
+
 # openpipelines 3.0.0
 
 ## BREAKING CHANGES